% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{.register_graph_context}
\alias{.register_graph_context}
\title{Register a graph context for the duration of the current frame}
\usage{
.register_graph_context(graph, free = FALSE, env = parent.frame())
}
\arguments{
\item{graph}{A \code{tbl_graph} object}

\item{free}{Should the active state of the graph be ignored?}

\item{env}{The environment where the context should be active}
}
\description{
This function sets the provided graph to be the context for tidygraph
algorithms, such as e.g. \code{\link[=node_is_center]{node_is_center()}}, for the duration of the current
environment. It automatically removes the graph once the environment exits.
}
\keyword{internal}
