% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{node_types}
\alias{node_types}
\alias{node_cut}
\alias{node_root}
\alias{node_leaf}
\alias{node_sink}
\alias{node_source}
\alias{node_isolated}
\alias{node_universal}
\alias{node_simplical}
\alias{node_center}
\title{Querying node types}
\usage{
node_cut()

node_root()

node_leaf()

node_sink()

node_source()

node_isolated()

node_universal(mode = "out")

node_simplical(mode = "out")

node_center(mode = "out")
}
\arguments{
\item{mode}{The way edges should be followed in the case of directed graphs.}
}
\value{
A logical vector of the same length as the number of nodes in the
graph.
}
\description{
These functions all lets the user query whether each node is of a certain
type. All of the functions returns a logical vector indicating whether the
node is of the type in question. Do note that the types are not mutually
exclusive and that nodes can thus be of multiple types.
}
\section{Functions}{
\itemize{
\item \code{node_cut}: is the node a cut node (articaultion node)

\item \code{node_root}: is the node a root in a tree

\item \code{node_leaf}: is the node a leaf in a tree

\item \code{node_sink}: does the node only have incomming edges

\item \code{node_source}: does the node only have outgoing edges

\item \code{node_isolated}: is the node unconnected

\item \code{node_universal}: is the node connected to all other nodes in the graph

\item \code{node_simplical}: are all the neighbors of the node connected

\item \code{node_center}: does the node have the minimal eccentricity in the graph
}}

