% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R, R/distinct.R, R/filter.R,
%   R/group_by.R, R/mutate.R, R/rename.R, R/sample_frac.R, R/sample_n.R,
%   R/select.R, R/slice.R, R/tibble.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{arrange}
\alias{reexports}
\alias{distinct}
\alias{reexports}
\alias{filter}
\alias{reexports}
\alias{top_n}
\alias{reexports}
\alias{group_by}
\alias{reexports}
\alias{ungroup}
\alias{reexports}
\alias{group_size}
\alias{reexports}
\alias{n_groups}
\alias{reexports}
\alias{groups}
\alias{reexports}
\alias{group_vars}
\alias{reexports}
\alias{mutate}
\alias{reexports}
\alias{transmute}
\alias{reexports}
\alias{mutate_all}
\alias{reexports}
\alias{mutate_at}
\alias{reexports}
\alias{rename}
\alias{reexports}
\alias{sample_frac}
\alias{reexports}
\alias{sample_n}
\alias{reexports}
\alias{select}
\alias{reexports}
\alias{contains}
\alias{reexports}
\alias{ends_with}
\alias{reexports}
\alias{everything}
\alias{reexports}
\alias{matches}
\alias{reexports}
\alias{num_range}
\alias{reexports}
\alias{one_of}
\alias{reexports}
\alias{starts_with}
\alias{reexports}
\alias{slice}
\alias{reexports}
\alias{as_tibble}
\alias{reexports}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{arrange}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{top_n}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{ungroup}}, \code{\link[dplyr]{group_size}}, \code{\link[dplyr]{n_groups}}, \code{\link[dplyr]{groups}}, \code{\link[dplyr]{group_vars}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{transmute}}, \code{\link[dplyr]{mutate_all}}, \code{\link[dplyr]{mutate_at}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{sample_frac}}, \code{\link[dplyr]{sample_n}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{contains}}, \code{\link[dplyr]{ends_with}}, \code{\link[dplyr]{everything}}, \code{\link[dplyr]{matches}}, \code{\link[dplyr]{num_range}}, \code{\link[dplyr]{one_of}}, \code{\link[dplyr]{starts_with}}, \code{\link[dplyr]{slice}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}}
}}

