% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidygraph-package.R
\docType{package}
\name{tidygraph-package}
\alias{tidygraph}
\alias{tidygraph-package}
\title{tidygraph: A Tidy API for Graph Manipulation}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

A graph, while not "tidy" in itself, can be thought of as two tidy
    data frames describing node and edge data respectively. 'tidygraph'
    provides an approach to manipulate these two virtual data frames using the
    API defined in the 'dplyr' package, as well as provides tidy interfaces to 
    a lot of common graph algorithms.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thomasp85/tidygraph}
  \item Report bugs at \url{https://github.com/thomasp85/tidygraph/issues}
}

}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomasp85@gmail.com}

}
