% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R, R/distinct.R, R/filter.R,
%   R/group_by.R, R/igraph.R, R/joins.R, R/mutate.R, R/pull.R, R/rename.R,
%   R/sample_frac.R, R/sample_n.R, R/select.R, R/slice.R, R/tbl_graph.R,
%   R/tibble.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{arrange}
\alias{distinct}
\alias{filter}
\alias{top_n}
\alias{group_by}
\alias{ungroup}
\alias{group_size}
\alias{n_groups}
\alias{groups}
\alias{group_vars}
\alias{group_data}
\alias{group_indices}
\alias{group_keys}
\alias{as.igraph}
\alias{left_join}
\alias{right_join}
\alias{inner_join}
\alias{full_join}
\alias{semi_join}
\alias{anti_join}
\alias{mutate}
\alias{transmute}
\alias{mutate_all}
\alias{mutate_at}
\alias{n}
\alias{pull}
\alias{rename}
\alias{sample_frac}
\alias{sample_n}
\alias{select}
\alias{contains}
\alias{ends_with}
\alias{everything}
\alias{matches}
\alias{num_range}
\alias{one_of}
\alias{starts_with}
\alias{slice}
\alias{tbl_vars}
\alias{as_tibble}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{anti_join}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{contains}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{ends_with}}, \code{\link[dplyr]{everything}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{full_join}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{group_data}}, \code{\link[dplyr]{group_indices}}, \code{\link[dplyr]{group_keys}}, \code{\link[dplyr]{group_size}}, \code{\link[dplyr]{group_vars}}, \code{\link[dplyr]{groups}}, \code{\link[dplyr]{inner_join}}, \code{\link[dplyr]{left_join}}, \code{\link[dplyr]{matches}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{mutate_all}}, \code{\link[dplyr]{mutate_at}}, \code{\link[dplyr]{n}}, \code{\link[dplyr]{n_groups}}, \code{\link[dplyr]{num_range}}, \code{\link[dplyr]{one_of}}, \code{\link[dplyr]{pull}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{right_join}}, \code{\link[dplyr]{sample_frac}}, \code{\link[dplyr]{sample_n}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{semi_join}}, \code{\link[dplyr]{slice}}, \code{\link[dplyr]{starts_with}}, \code{\link[dplyr]{tbl_vars}}, \code{\link[dplyr]{top_n}}, \code{\link[dplyr]{transmute}}, \code{\link[dplyr]{ungroup}}}

  \item{igraph}{\code{\link[igraph]{as.igraph}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}}
}}

