% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R, R/distinct.R, R/filter.R,
%   R/group_by.R, R/igraph.R, R/joins.R, R/mutate.R, R/pull.R, R/rename.R,
%   R/sample_frac.R, R/sample_n.R, R/select.R, R/slice.R, R/tbl_graph.R,
%   R/tibble.R, R/tidyr-utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{arrange}
\alias{distinct}
\alias{filter}
\alias{top_n}
\alias{group_by}
\alias{ungroup}
\alias{group_size}
\alias{n_groups}
\alias{groups}
\alias{group_vars}
\alias{group_data}
\alias{group_indices}
\alias{group_keys}
\alias{as.igraph}
\alias{left_join}
\alias{right_join}
\alias{inner_join}
\alias{full_join}
\alias{semi_join}
\alias{anti_join}
\alias{mutate}
\alias{transmute}
\alias{mutate_all}
\alias{mutate_at}
\alias{n}
\alias{pull}
\alias{rename}
\alias{sample_frac}
\alias{sample_n}
\alias{select}
\alias{contains}
\alias{ends_with}
\alias{everything}
\alias{matches}
\alias{num_range}
\alias{one_of}
\alias{starts_with}
\alias{slice}
\alias{slice_head}
\alias{slice_tail}
\alias{slice_min}
\alias{slice_max}
\alias{slice_sample}
\alias{tbl_vars}
\alias{as_tibble}
\alias{\%>\%}
\alias{replace_na}
\alias{drop_na}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:filter-joins]{anti_join}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr:reexports]{contains}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr:reexports]{ends_with}}, \code{\link[dplyr:reexports]{everything}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr:mutate-joins]{full_join}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{group_data}}, \code{\link[dplyr:group_data]{group_indices}}, \code{\link[dplyr:group_data]{group_keys}}, \code{\link[dplyr:group_data]{group_size}}, \code{\link[dplyr:group_data]{group_vars}}, \code{\link[dplyr:group_data]{groups}}, \code{\link[dplyr:mutate-joins]{inner_join}}, \code{\link[dplyr:mutate-joins]{left_join}}, \code{\link[dplyr:reexports]{matches}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{mutate_all}}, \code{\link[dplyr:mutate_all]{mutate_at}}, \code{\link[dplyr:context]{n}}, \code{\link[dplyr:group_data]{n_groups}}, \code{\link[dplyr:reexports]{num_range}}, \code{\link[dplyr:reexports]{one_of}}, \code{\link[dplyr]{pull}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr:mutate-joins]{right_join}}, \code{\link[dplyr:sample_n]{sample_frac}}, \code{\link[dplyr]{sample_n}}, \code{\link[dplyr]{select}}, \code{\link[dplyr:filter-joins]{semi_join}}, \code{\link[dplyr]{slice}}, \code{\link[dplyr:slice]{slice_head}}, \code{\link[dplyr:slice]{slice_max}}, \code{\link[dplyr:slice]{slice_min}}, \code{\link[dplyr:slice]{slice_sample}}, \code{\link[dplyr:slice]{slice_tail}}, \code{\link[dplyr:reexports]{starts_with}}, \code{\link[dplyr]{tbl_vars}}, \code{\link[dplyr]{top_n}}, \code{\link[dplyr]{transmute}}, \code{\link[dplyr:group_by]{ungroup}}}

  \item{igraph}{\code{\link[igraph]{as.igraph}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}}

  \item{tidyr}{\code{\link[tidyr]{drop_na}}, \code{\link[tidyr]{replace_na}}}
}}

